+++ ftlib/consensus/gossip/__init__.py
@@ -1,1 +1,1 @@
-from ftlib.consensus.gossip.impl import Gossip
+from ftlib.consensus.gossip.impl import Gossip  # noqa: F401
