+++ ftlib/consensus/shared_storage/__init__.py
@@ -1,1 +1,1 @@
-from ftlib.consensus.shared_storage.impl import SharedStorage
+from ftlib.consensus.shared_storage.impl import SharedStorage  # noqa: F401
