+++ ftlib/framework/__init__.py
@@ -1,0 +1,1 @@
+from ftlib.framework import dummy_nccl, pytorch  # noqa: F401
