+++ requirements.txt
@@ -1,3 +1,4 @@
 grpcio==1.24.1
 grpcio-tools==1.24.1
 numpy
+torch==1.2.0
